Fix tests after moving functions.
